solalendar-v2/requirements.txt
streamlit
kerykeion
lunar-python
pytz
pyswisseph